feat: add hook for blogposts
